[S10P12A306-679] refactor: Add security validation

+++ BackEnd/bid/bid-admin/src/main/java/com/ssafy/bid/domain/user/service/UserServiceImpl.java
@@ -11,6 +11,7 @@ import com.ssafy.bid.domain.user.Admin;
 import com.ssafy.bid.domain.user.Student;
 import com.ssafy.bid.domain.user.TelAuthentication;
 import com.ssafy.bid.domain.user.User;
+import com.ssafy.bid.domain.user.UserType;
 import com.ssafy.bid.domain.user.dto.AdminPasswordUpdateRequest;
 import com.ssafy.bid.domain.user.dto.AdminSaveRequest;
 import com.ssafy.bid.domain.user.dto.BallsFindResponse;
@@ -25,6 +26,7 @@ import com.ssafy.bid.domain.user.dto.UserUpdateRequest;
 import com.ssafy.bid.domain.user.repository.TelAuthenticationRepository;
 import com.ssafy.bid.domain.user.repository.UserRepository;
 import com.ssafy.bid.global.error.exception.AuthenticationFailedException;
+import com.ssafy.bid.global.error.exception.AuthorizationFailedException;
 import com.ssafy.bid.global.error.exception.InvalidParameterException;
 import com.ssafy.bid.global.error.exception.ResourceAlreadyExistsException;
 import com.ssafy.bid.global.error.exception.ResourceNotFoundException;
@@ -48,7 +50,7 @@ public class UserServiceImpl implements UserService {
 	) {
 		boolean exists = userRepository.existsByIdAndTel(request.getId(), request.getTel());
 		if (!exists) {
-			throw new ResourceNotFoundException("해당 id와 tel을 가진 User 엔티티가 없음.", request.getTel());
+			throw new ResourceNotFoundException("인증코드전송: 해당 id와 tel 을 가진 User 엔티티가 없음.", request.getTel());
 		}
 
 		TelAuthentication telAuthentication = createAndSendAuthenticationCode(request.getTel());
@@ -81,11 +83,11 @@ public class UserServiceImpl implements UserService {
 	@Transactional
 	public boolean checkTelAuthentication(TelAuthenticationCheckRequest request) {
 		TelAuthentication telAuthentication = telAuthenticationRepository.findById(request.getTel())
-			.orElseThrow(() -> new InvalidParameterException("해당 번호로 전송된 인증코드가 만료됐거나 없음.", request.getTel()));
+			.orElseThrow(() -> new InvalidParameterException("인증코드확인: 해당 번호로 전송된 인증코드가 만료됐거나 없음.", request.getTel()));
 
 		boolean equals = telAuthentication.getCode().equals(request.getCode());
 		if (!equals) {
-			new AuthenticationFailedException("인증코드 불일치.");
+			throw new AuthenticationFailedException("인증코드확인: 인증코드 불일치.");
 		}
 
 		telAuthentication.authenticate();
@@ -93,7 +95,7 @@ public class UserServiceImpl implements UserService {
 		return equals;
 	}
 
-	@Override //TODO: 아이디 중복여부 체크 인덱스 조회로 수정
+	@Override
 	public boolean isDuplicated(String id) {
 		return userRepository.existsById(id);
 	}
@@ -107,13 +109,13 @@ public class UserServiceImpl implements UserService {
 	@Transactional
 	public void saveAdmin(AdminSaveRequest request) {
 		if (!request.getPassword().equals(request.getConfirmPassword())) {
-			throw new InvalidParameterException("패스워드와 패스워드 확인 불일치..", request.getPassword());
+			throw new InvalidParameterException("회원가입: 패스워드와 패스워드 확인 불일치.", request.getPassword());
 		}
 
 		TelAuthentication telAuthentication = telAuthenticationRepository.findById(request.getTel())
-			.orElseThrow(() -> new AuthenticationFailedException("인증되지 않은 회원임."));
+			.orElseThrow(() -> new AuthenticationFailedException("회원가입: 인증되지 않은 회원임."));
 		if (!telAuthentication.getIsAuthenticated()) {
-			throw new AuthenticationFailedException("인증되지 않은 회원임.");
+			throw new AuthenticationFailedException("회원가입: 인증되지 않은 회원임.");
 		}
 
 		Admin admin = request.toEntity(passwordEncoder);
@@ -122,19 +124,29 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	@Transactional
-	public void saveStudent(StudentSaveRequest request) {
-		if (userRepository.existsById(request.getId())) {
-			throw new ResourceAlreadyExistsException("회원가입하려는 아이디가 중복됨.", request.getId());
+	public void saveStudent(UserType userType, StudentSaveRequest request) {
+		if (!userType.equals(UserType.ADMIN)) {
+			throw new AuthorizationFailedException("학생등록: Admin 권한 사용자가 아님.");
 		}
+
+		if (userRepository.existsById(request.getId())) {
+			throw new ResourceAlreadyExistsException("학생등록: 회원가입하려는 아이디가 중복됨.", request.getId());
+		}
+
 		Student student = request.toEntity(passwordEncoder);
 		userRepository.save(student);
 	}
 
 	@Override
 	@Transactional
-	public void resetStudentPassword(int userNo) {
+	public void resetStudentPassword(UserType userType, int userNo) {
+		if (!userType.equals(UserType.ADMIN)) {
+			throw new AuthorizationFailedException("학생패스워드초기화: Admin 권한 사용자가 아님.");
+		}
+
 		Student student = userRepository.findStudentByUserNo(userNo)
-			.orElseThrow(() -> new ResourceNotFoundException("패스워드 초기화하려는 User 엔티티가 없음.", userNo));
+			.orElseThrow(() -> new ResourceNotFoundException("학생패스워드초기화: 패스워드 초기화하려는 Student 엔티티가 없음.", userNo));
+
 		student.resetPassword(passwordEncoder);
 	}
 
@@ -142,72 +154,96 @@ public class UserServiceImpl implements UserService {
 	@Transactional
 	public void updateAdminPassword(AdminPasswordUpdateRequest request) {
 		Admin admin = userRepository.findAdminById(request.getId())
-			.orElseThrow(() -> new ResourceNotFoundException("패스워드 수정하려는 User 엔티티가 없음.", request.getId()));
+			.orElseThrow(() -> new ResourceNotFoundException("관리자패스워드변경: 패스워드 수정하려는 Admin 엔티티가 없음.", request.getId()));
 
 		if (!request.getNewPassword().equals(request.getNewPasswordCheck())) {
-			throw new InvalidParameterException("패스워드와 패스워드 확인 불일치.", request.getNewPassword());
+			throw new InvalidParameterException("관리자패스워드변경: 패스워드와 패스워드 확인 불일치.", request.getNewPassword());
 		}
 
 		TelAuthentication telAuthentication = telAuthenticationRepository.findById(admin.getTel())
-			.orElseThrow(() -> new AuthenticationFailedException("인증되지 않은 회원임."));
+			.orElseThrow(() -> new AuthenticationFailedException("관리자패스워드변경: 인증되지 않은 회원임."));
 		if (!telAuthentication.getIsAuthenticated()) {
-			throw new AuthenticationFailedException("인증되지 않은 회원임.");
+			throw new AuthenticationFailedException("관리자패스워드변경: 인증되지 않은 회원임.");
 		}
 
 		admin.changePassword(passwordEncoder.encode(request.getNewPassword()));
 	}
 
 	@Override
-	public List<StudentsGetResponse> getStudents(int gradeNo) {
+	public List<StudentsGetResponse> getStudents(UserType userType, int gradeNo) {
+		if (!userType.equals(UserType.ADMIN)) {
+			throw new AuthorizationFailedException("학생목록조회: Admin 권한 사용자가 아님.");
+		}
+
 		List<StudentsGetResponse> responses = userRepository.findAllStudentByGradeNo(gradeNo);
 		if (responses.isEmpty()) {
-			throw new ResourceNotFoundException("조회하려는 User 가 없음.", gradeNo);
+			throw new ResourceNotFoundException("학생목록조회: 조회하려는 Student 가 없음.", gradeNo);
 		}
+
 		return responses;
 	}
 
 	@Override
 	public String getUserId(UserIdFindRequest request) {
 		return userRepository.findUserIdByNameAndTel(request.getName(), request.getTel())
-			.orElseThrow(() -> new ResourceNotFoundException("조회하려는 User 가 없음.", request.getName()));
+			.orElseThrow(() -> new ResourceNotFoundException("관리자ID조회: 조회하려는 Admin 가 없음.", request.getName()));
 	}
 
 	@Override
 	@Transactional
-	public void updateUser(Integer userNo, UserUpdateRequest request) {
+	public void updateUser(UserType userType, int userNo, UserUpdateRequest request) {
+		if (!userType.equals(UserType.ADMIN)) {
+			throw new AuthorizationFailedException("관리자수정: Admin 권한 사용자가 아님.");
+		}
+
 		Admin admin = userRepository.findAdminByUserNo(userNo)
-			.orElseThrow(() -> new ResourceNotFoundException("수정하려는 User 가 없음.", userNo));
+			.orElseThrow(() -> new ResourceNotFoundException("관리자수정: 수정하려는 Admin 가 없음.", userNo));
 
 		admin.update(request.getName(), request.getSchoolNo(), request.getTel());
 	}
 
 	@Override
 	@Transactional
-	public void deleteUser(Integer userNo, UserDeleteRequest request) {
+	public void deleteUser(UserType userType, int userNo, UserDeleteRequest request) {
+		if (!userType.equals(UserType.ADMIN)) {
+			throw new AuthorizationFailedException("관리자탈퇴: Admin 권한 사용자가 아님.");
+		}
+
 		User user = userRepository.findById(userNo)
-			.orElseThrow(() -> new ResourceNotFoundException("탈퇴하려는 User 가 없음.", userNo));
+			.orElseThrow(() -> new ResourceNotFoundException("관리자탈퇴: 탈퇴하려는 Admin 가 없음.", userNo));
+
 		if (!passwordEncoder.matches(request.getPassword(), user.getPassword())) {
-			throw new InvalidParameterException("패스워드와 패스워드 확인 불일치.", request.getPassword());
+			throw new InvalidParameterException("관리자탈퇴: 패스워드와 패스워드 확인 불일치.", request.getPassword());
 		}
 		userRepository.delete(user);
 	}
 
 	@Override
-	public List<BallsFindResponse> getAllBalls(int gradeNo) {
+	public List<BallsFindResponse> getAllBalls(UserType userType, int gradeNo) {
+		if (!userType.equals(UserType.ADMIN)) {
+			throw new AuthorizationFailedException("대포목록조회: Admin 권한 사용자가 아님.");
+		}
+
 		List<BallsFindResponse> responses = userRepository.findAllBallsByGradeNo(gradeNo);
 		if (responses.isEmpty()) {
-			throw new ResourceNotFoundException("조회하려는 Grade 가 없음.", gradeNo);
+			throw new ResourceNotFoundException("대포목록조회: 조회하려는 Grade 가 없음.", gradeNo);
 		}
+
 		return userRepository.findAllBallsByGradeNo(gradeNo);
 	}
 
 	@Override
 	@Transactional
-	public void resetAllBalls(int gradeNo) {
+	public void resetAllBalls(UserType userType, int gradeNo) {
+		if (!userType.equals(UserType.ADMIN)) {
+			throw new AuthorizationFailedException("대포초기화: Admin 권한 사용자가 아님.");
+		}
+
 		List<Student> responses = userRepository.findAllStudentsByGradeNo(gradeNo);
 		if (responses.isEmpty()) {
-			throw new ResourceNotFoundException("조회하려는 User 가 없음.", gradeNo);
+			throw new ResourceNotFoundException("대포초기화: 조회하려는 Student 가 없음.", gradeNo);
 		}
+
 		responses.forEach(Student::resetBalls);
 	}
 }
